requirements.txt
dmlab2d>=1.0.0
dm-meltingpot>=2.2.0
dm-env>=1.6
numpy>=1.24.0
scipy>=1.10.0
matplotlib>=3.7.0
rich>=13.0.0
pytest>=7.0.0
pandas>=2.0.0
